# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hy3-ContestLens",
    "description": "Hy3 process evaluation and error localization for NOIP 2018",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/problems/{problem_id}": {
      "get": {
        "summary": "Get Problem",
        "operationId": "get_problem_api_v1_problems__problem_id__get",
        "parameters": [
          {
            "name": "problem_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Problem Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Problem Api V1 Problems  Problem Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/problems/{problem_id}/tests/summary": {
      "get": {
        "summary": "Test Summary",
        "operationId": "test_summary_api_v1_problems__problem_id__tests_summary_get",
        "parameters": [
          {
            "name": "problem_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Problem Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Test Summary Api V1 Problems  Problem Id  Tests Summary Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resource-scopes:validate": {
      "post": {
        "summary": "Validate Scope",
        "operationId": "validate_scope_api_v1_resource_scopes_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidatePathRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Validate Scope Api V1 Resource Scopes Validate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resource-scopes": {
      "get": {
        "summary": "List Scopes",
        "operationId": "list_scopes_api_v1_resource_scopes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response List Scopes Api V1 Resource Scopes Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Grant Scope",
        "operationId": "grant_scope_api_v1_resource_scopes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantPathRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Grant Scope Api V1 Resource Scopes Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resource-scopes/{scope_id}": {
      "get": {
        "summary": "Inspect Scope",
        "operationId": "inspect_scope_api_v1_resource_scopes__scope_id__get",
        "parameters": [
          {
            "name": "scope_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scope Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Inspect Scope Api V1 Resource Scopes  Scope Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resource-scopes/{scope_id}:discover": {
      "post": {
        "summary": "Discover",
        "operationId": "discover_api_v1_resource_scopes__scope_id__discover_post",
        "parameters": [
          {
            "name": "scope_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scope Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscoverRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Discover Api V1 Resource Scopes  Scope Id  Discover Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/problems/{problem_id}/resource-binding": {
      "post": {
        "summary": "Bind",
        "operationId": "bind_api_v1_problems__problem_id__resource_binding_post",
        "parameters": [
          {
            "name": "problem_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Problem Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Bind Api V1 Problems  Problem Id  Resource Binding Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get Binding",
        "operationId": "get_binding_api_v1_problems__problem_id__resource_binding_get",
        "parameters": [
          {
            "name": "problem_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Problem Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Binding Api V1 Problems  Problem Id  Resource Binding Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/runs/{run_id}": {
      "get": {
        "summary": "Get Run",
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Run Api V1 Runs  Run Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BindRequest": {
        "properties": {
          "scope_id": {
            "type": "string",
            "title": "Scope Id"
          },
          "candidate_id": {
            "type": "string",
            "title": "Candidate Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "scope_id",
          "candidate_id"
        ],
        "title": "BindRequest"
      },
...
      "GrantPathRequest": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "confirmed": {
            "type": "boolean",
            "title": "Confirmed"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "path",
          "confirmed"
        ],
        "title": "GrantPathRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidatePathRequest": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "path"
        ],
        "title": "ValidatePathRequest"
      },
...
    }
  }
}